+++ src/Products/MeetingCommunes/tests/MeetingCommunesTestCase.py
@@ -11,9 +11,6 @@ from Products.PloneMeeting.tests.PloneMeetingTestCase import PloneMeetingTestCas
 class MeetingCommunesTestCase(PloneMeetingTestCase, MeetingCommunesTestingHelpers):
     """Base class for defining MeetingCommunes test cases."""
 
-    # Some default content
-    descriptionText = '<p>Some description</p>'
-    decisionText = '<p>Some decision.</p>'
     # by default, PloneMeeting's test file testPerformances.py and
     # testConversionWithDocumentViewer.py' are ignored, override the subproductIgnoredTestFiles
     # attribute to take these files into account